Move view, add search
