+++ app/views/moderatorIndex.pug
@@ -18,6 +18,8 @@ block navbar
       li
         a(href='/moderator/change_card') Mistet kort
       li
+        a(href='/moderator/manage_register') Register
+      li
         a(href='/moderator/deactivate_users') Deaktiver brukere
 
 block content
