+++ client/controllers/deactivateUsersCtrl.js
@@ -4,22 +4,8 @@ module.exports = [
   '$scope',
   '$route',
   'userService',
-  'registerService',
   'alertService',
-  function ($scope, $route, userService, registerService, alertService) {
-    $scope.registers = [];
-    function getRegisterEntries() {
-      registerService.getRegisterEntries().then(
-        function (response) {
-          $scope.registers = response.data;
-        },
-        function (response) {
-          alertService.addError(response.message);
-        }
-      );
-    }
-    getRegisterEntries();
-
+  function ($scope, $route, userService, alertService) {
     $scope.deactivateNonAdminUsers = function () {
       userService.deactivateNonAdminUsers().then(
         function (response) {
@@ -32,20 +18,6 @@ module.exports = [
       );
 
       $route.reload();
-    };
-
-    $scope.deleteRegister = function (register) {
-      if (confirm('Er du sikker på at du vil slette denne brukeren?')) {
-        registerService.deleteRegisterEntry(register).then(
-          function (response) {
-            alertService.addSuccess(response.data.message);
-            $route.reload();
-          },
-          function (response) {
-            alertService.addError(response.message);
-          }
-        );
-      }
     };
   },
 ];
