+++ client/controllers/index.js
@@ -12,4 +12,5 @@ angular
   .controller('logoutController', require('./logoutCtrl'))
   .controller('retrieveVoteController', require('./retrieveVoteCtrl'))
   .controller('showQRController', require('./showQRCtrl'))
-  .controller('toggleUserController', require('./toggleUserCtrl'));
+  .controller('toggleUserController', require('./toggleUserCtrl'))
+  .controller('manageRegisterController', require('./manageRegisterCtrl'));
