+++ client/styles/main.styl
@@ -49,11 +49,6 @@ header
                 &:focus
                     text-decoration none
 
-thead
-  font-style bold
-  border-bottom 1px solid black
-
-
 .header
     text-align center
     font-size 60px
